move array utils to helper
