+++ functions/src/helper/array.ts
